.builds: disable OpenGL tests on FreeBSD
For some reason, the Mesa OpenGL software renderer fails to correctly
render some of the tests in gpu/internal/rendertest. The Linux builder
uses the Vulkan software renderer, so only the FreeBSD builders fails.

The root cause may be a Gio issue, but since all the other backends work,
including hardware OpenGL, there is a likelyhood the issue is in Mesa.
I don't have the temperament to debug Mesa issues, so this change disables
software OpenGL on FreeBSD, which will cause the rendertests to skip.

Running Vulkan tests under FreeBSD is blocked on

https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=257303

In practice it won't matter: FreeBSD and Linux use the same software
for software renderers.

Signed-off-by: Elias Naur <mail@eliasnaur.com>