flake.lock: [NixOS] delete to avoid GPU driver version mismatch
A `flake.lock` file pins the versions of packages Gio requires to build.
However, pinning the `libGL` package may result in a runtime incompatibility
between the user space GPU driver and the system GPU driver. The result is
an error such as

	glGetDisplay failed: 0x300

I don't know a way to keep the convenience of pinned flake.lock versions
combined with a system-defined `libGL` package version. This change implements
a workaround: exclude `flake.lock` from version control and let the first
use of `flake.nix` auto-create a lock file that (hopefully) is compatible
with the system.

References: https://todo.sr.ht/~eliasnaur/gio/417
Signed-off-by: Elias Naur <mail@eliasnaur.com>