text: eliminate path cache memory leak
This commit alters the method we use to check for valid cache hits
in the text path cache. Previously we stored the entire text.Layout
that was provided when the cache entry was set so that we could ensure
only identical text.Layouts would produce hits (guarding against hash
collisions). This commit instead pulls the glyph IDs for every glyph
in the text.Layout and stores them in the cache. This uses far less
memory and seems to allow cache entries to be GCed after eviction.

Fixes: https://todo.sr.ht/~eliasnaur/gio/418
Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>